Refactor executor to runtime, compiler to executor, parser to compiler
