+++ src/go-virtual-machine-main/virtual-machine/compiler/golang_parser.ts
