+++ src/go-virtual-machine-main/virtual-machine/compiler/parser.ts
