+++ src/go-virtual-machine-main/virtual-machine/compiler/tokens.ts
