+++ src/go-virtual-machine-main/virtual-machine/executor/environment.ts
