+++ src/go-virtual-machine-main/virtual-machine/executor/instructions/concurrent.ts
@@ -1,4 +1,4 @@
-import { Process } from '../../executor/process'
+import { Process } from '../../runtime/process'
 import { ArrayNode } from '../../heap/types/array'
 import {
   ChannelNode,
