+++ src/go-virtual-machine-main/virtual-machine/executor/instructions/funcs.ts
@@ -1,4 +1,4 @@
-import { Process } from '../../executor/process'
+import { Process } from '../../runtime/process'
 import {
   CallRefNode,
   DeferFuncNode,
