+++ src/go-virtual-machine-main/virtual-machine/executor/instructions/index.ts
