+++ src/go-virtual-machine-main/virtual-machine/executor/instructions/store.ts
@@ -1,4 +1,4 @@
-import { Process } from '../../executor/process'
+import { Process } from '../../runtime/process'
 
 import { Instruction } from './base'
