+++ src/go-virtual-machine-main/virtual-machine/executor/typing/type_environment.ts
