+++ src/go-virtual-machine-main/virtual-machine/runtime/ops.ts
